migrando metodos para graficos a graphics.py
