+++ core/charts/ep.py
@@ -1,12 +1,8 @@
 # -*- coding: utf-8 -*-
 
-import json
-
-
 from django.db import connection
 from django.db.models import Sum
-from django.shortcuts import render_to_response, render
-from django.template import RequestContext
+from django.shortcuts import render
 
 from chartit import Chart, RawDataPool
